Version number link label is incorrect
The v2.5 version of this page states that the user will see a link labeled "v2.6" in the Rancher UI. This is clearly a typo. If the user is on 2.5.x, they'll see v2.5.x listed in that link. The link points to the /dashboard/about page in the Rancher UI which contains further information about the user's version of Rancher.