update to mention versioning in docs readme
We sometimes get questions about how to contribute. We have a lot of information on internal sites but it isn't very easy to find. We might want to make more of that information public on Github, as it makes it easier to collaborate if it's in such a visible place. Adding some simplified info about versioning as it comes up frequently.