Convert all helm-template instructions to helm-install (#828)
* Adjust command to install cert-manager

The directory `cert-manager` is not created when using `helm install`
instead of `helm template --output-dir .`.

Also, the CRDs are downloaded in the current directory, so adjust the
commands as well.

* Convert helm template to helm upgrade/install

Since we're using Helm hooks, we cannot use `helm template`, otherwise
things will break. The commands are updated to use `helm install` and
`helm upgrade`.