Document workaround for migrating across distributions
Migrating rancher across different Kubernetes distributions causes the local
cluster object not to specify the correct distribution and driver.

This commit adds a workaround to be applied after restoration for a migration
and before bringing Rancher up, so that the local cluster will have the
correct data for the new distribution it runs on.