Fix v1.29 in-tree provider usage instructions for Azure (#2033)
* Update azure.md for in-tree cloud providers in v1.29

Correct 'DisableKubeletCloudCredentialProvider' feature gate name to 'DisableKubeletCloudCredentialProviders'. Add both feature gates in the argument for Kubelet, Controller Manager and API Server since both are required in all of these components in v1.29. Also added the Cloud Provider Integration changes blog post for more details along with an example of the argument.

* Add v1.29 Azure in-tree provider details on versioned docs compatible with v1.29

* Apply new link to upstream documentation

Apply new link to upstream documentation for the Azure in-tree cloud providers.

Co-authored-by: Lucas Saintarbor <lucas.saintarbor@suse.com>

* Added Azure in-tree provider corrections to 2.13 docs

---------

Co-authored-by: Lucas Saintarbor <lucas.saintarbor@suse.com>